requirements.txt
fastapi
python-pptx
pydantic
uvicorn
g4f
PyMuPDF
python-dotenv
python-multipart
python-jose
aiofiles
requests
reportlab
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
orjson==3.10.12
